requirements.txt
Flask
numpy
pillow
torch
torchvision
